requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
itsdangerous==0.24
lxml==3.4.0
python-dateutil==2.2
requests==2.4.3
six==1.8.0
wsgiref==0.1.2
beautifulsoup4==4.3.2
